+++ telegram/error.py
@@ -18,7 +18,7 @@
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
 # pylint: disable=C0115
 """This module contains an object that represents Telegram errors."""
-from typing import Tuple
+from typing import Tuple, Union
 
 
 def _lstrip_str(in_s: str, lstr: str) -> str:
@@ -149,3 +149,16 @@ class Conflict(TelegramError):
 
     def __reduce__(self) -> Tuple[type, Tuple[str]]:
         return self.__class__, (self.message,)
+
+
+class TelegramDecryptionError(TelegramError):
+    """Something went wrong with decryption."""
+
+    __slots__ = ('_msg',)
+
+    def __init__(self, message: Union[str, Exception]):
+        super().__init__(f"TelegramDecryptionError: {message}")
+        self._msg = str(message)
+
+    def __reduce__(self) -> Tuple[type, Tuple[str]]:
+        return self.__class__, (self._msg,)
